# directory: optimize parallel deserialization in `Directory::load`