I am not sure why, but it seems that using `.into_par_iter()` on the entries vector is not correctly optimized by the compiler in this case. Unnecessary moves led to significant overhead, which would cancel out the benefit of parallelization on large directories.

#### **Performance of `Directory::load`**
observed on 32_000_000 entries, with 144 cores  
  
||Before<br />`.into_par_iter()`|After<br />`.par_iter()`|
|-|-|-|
|open DB, read capacity|3s|3s|
|load `entries` Vec|33s|33s|
|deserialize (in parallel)|**33s**|**2s**|
|writing `keycards` Vec|7s|7s|
|**Total**|**76s**|**45s**|